patch/cell: rework handling of adjacencies and interfaces
When creating a new cell, it is possible to decide whether the
information about the neighbourhood (adjacencies and interfaces)
will be stored. If the neighbourhood data are stored, there will
always be at least one adjacency/interface entry for each face.
If a face is not linked with an interface/neighbout, its entry
needs to be set to the placeholder value NULL_ID. When multiple
interfaces/neighbours are linked to a face, all entries must point
to valid interfaces/negihbours.